v2/requirements.txt
streamlit>=1.36.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
elevenlabs>=1.0.0
ffmpeg-python>=0.2.0
moviepy>=2.0.0
python-dotenv>=1.0.0
aiolimiter>=1.1.0
requests>=2.31.0
Pillow>=10.0.0
sqlalchemy>=2.0.0
tenacity>=8.0.0
weasyprint>=62.0
markdown>=3.5
Jinja2>=3.1.0
python-docx>=1.0.0
tavily-python>=0.5.0
